+++ tests/test_command_modbus.py
@@ -27,6 +27,19 @@ async def test_command_modbus_read(actor: ECPActor):
     assert read_cmd.replies.get("register")["value"]
 
 
+async def test_command_modbus_read_override(actor: ECPActor, mocker: MockerFixture):
+    mocker.patch.dict(
+        actor.plc.modbus.config,
+        {"overrides": {"oxygen_read_utilities_room": 25.0}},
+    )
+
+    read_cmd = await actor.invoke_mock_command("modbus read oxygen_read_utilities_room")
+    await read_cmd
+
+    assert read_cmd.status.did_succeed
+    assert read_cmd.replies.get("register")["value"] == 25.0
+
+
 async def test_command_modbus_read_address(actor: ECPActor):
     read_cmd = await actor.invoke_mock_command("modbus read --register-type coil 1")
     await read_cmd